integration-Review Product Wishlist User complete

+++ agricolae/app/Http/Controllers/ReviewController.php
@@ -5,7 +5,9 @@ namespace App\Http\Controllers;
 use App\Product;
 use Illuminate\Http\Request;
 use App\Review;
+use App\User;
 use Illuminate\Cache\RedisTaggedCache;
+use Illuminate\Support\Facades\Auth;
 
 //use App\User;
 
@@ -17,7 +19,12 @@ class ReviewController extends Controller
         $data = []; //to be sent to the view
         $review = Review::findOrFail($id);
         $product = Product::findOrFail($review->getProductId());
-        //$user = User::findOrFail($review->getUserId());
+        $user = User::findOrFail(Auth::user()->id);
+
+        if ($review->getUserId() != $user->getId())
+        {
+            return redirect()->route('home.index')->with('delted' ,"You cannot access this site"); 
+        }
         
         $data["title"] = "Review";
         $data["review"] = $review;
@@ -31,7 +38,9 @@ class ReviewController extends Controller
     {
         $data = []; //to be sent to the view
         $data["title"] = "Reviews";
-        $data["reviews"] = Review::all()->sortByDesc('id');
+        $user = User::findOrFail(Auth::user()->id);
+        $data["reviews"] = Review::where('user_id', $user->id)->get()->sortByDesc('id');
+        
 
         return view('review.list')->with("data",$data);
     }
@@ -49,12 +58,15 @@ class ReviewController extends Controller
     {
 
         $request->validate(Review::validateRules());
+        $user = User::findOrFail(Auth::user()->id);
 
         $review = new Review;
+        $review->user_id = $user->getId();
+        $review->product_id = $product->getId();
         $review->title = $request["title"];
         $review->description = $request["description"];
         $review->score = $request["score"];
-        $review->product_id = $product->getId();
+        
         $review->save();
 
         return redirect()->route('product.show' ,$product->getId());
@@ -64,8 +76,15 @@ class ReviewController extends Controller
     {
         $data = [];
         $data["title"] = "Edit Review";
-        
         $review = Review::findOrFail($id);
+
+        $user = User::findOrFail(Auth::user()->id);
+
+        if ($review->getUserId() != $user->getId())
+        {
+            return redirect()->route('home.index')->with('delted' ,"You cannot access this site"); 
+        }
+
         $data['review'] = $review;
 
         return view('review.edit')->with(["data" => $data]);
@@ -74,6 +93,14 @@ class ReviewController extends Controller
     public function update(Request $request, $id)
     {
         $review = Review::findOrFail($id);
+
+        $user = User::findOrFail(Auth::user()->id);
+
+        if ($review->getUserId() != $user->getId())
+        {
+            return redirect()->route('home.index')->with('delted' ,"You cannot access this site"); 
+        }
+
         $validate = $request->validate(Review::validateRules());
 
         if (!$validate)
@@ -83,12 +110,20 @@ class ReviewController extends Controller
 
         $review->update($request->only(["title","description","score"]));
         
-        return redirect()->route('review.show', [$id, $review->product_id]);
+        return redirect()->route('review.show', [$id, $review->product_id])->with('success', 'Review has been succesfully edited');
 
     }
 
     public function delete($id){
         $review = Review::find($id);
+
+        $user = User::findOrFail(Auth::user()->id);
+
+        if ($review->getUserId() != $user->getId())
+        {
+            return redirect()->route('home.index')->with('delted' ,"You cannot access this site"); 
+        }
+
         $review->delete();
         return redirect()->route('review.list')->with('deleted', 'Review has been deleted!');
     }
