+++ agricolae/app/Http/Controllers/WishlistController.php
@@ -5,7 +5,9 @@ namespace App\Http\Controllers;
 use App\Wishlist;
 use Illuminate\Http\Request;
 use App\Product;
+use App\User;
 use Illuminate\Cache\RedisTaggedCache;
+use Illuminate\Support\Facades\Auth;
 
 class WishlistController extends Controller
 {
@@ -13,8 +15,9 @@ class WishlistController extends Controller
     public function list()
     {
         $data = []; //to be sent to the view
-        $data["title"] = 'WishList';
-        $data["wishlists"] = WishList::all()->sortByDesc('id');
+        $data["title"] = 'Wishlist';
+        $user = User::findOrFail(Auth::user()->id);
+        $data["wishlists"] = Wishlist::where('user_id', $user->id)->get()->sortByDesc('id');
 
         return view('wishlist.list')->with("data",$data);
     }
@@ -22,18 +25,28 @@ class WishlistController extends Controller
     public function save(Request $request, Product $product)
     {
 
-        $request->validate(WishList::validateRules());
+        $request->validate(Wishlist::validateRules());
+        $user = User::findOrFail(Auth::user()->id);
 
-        $wishlist = new WishList;
-        $wishlist->title = $request["title"];
+        $wishlist = new Wishlist;
+        $wishlist->title = $request['title'];
         $wishlist->product_id = $product->getId();
+        $wishlist->user_id = $user->getId();
         $wishlist->save();
 
         return redirect()->route('wishlist.list');
     }
 
     public function delete($id){
-        $wishlist = WishList::find($id);
+        $wishlist = Wishlist::find($id);
+
+        $user = User::findOrFail(Auth::user()->id);
+
+        if ($wishlist->getUserId() != $user->getId())
+        {
+            return redirect()->route('home.index')->with('delted' ,"You cannot access this site"); 
+        }
+
         $wishlist->delete();
         return redirect()->route('wishlist.list')->with('deleted', 'Wish List has been deleted!');
     }
